NR.m: Fix rounding error in AlignPhases
The algorithm was sensitive to phases which were distributed around a multiple of Pi at the reference point.  This would be very rare, which is why this has not been noticed before.  Given a set of phases near a multiple of Pi, the 2 Pi addition could have been computed by either rounding up or down, causing a spread of +- 2 Pi in the returned phases.  If these were used in extrapolation, the result would be nonsense (though it would have been caught by estimates of extrapolation error).  The algorithm has been changed to use the first phase, rather than 0, as the reference, so as long as the phases at the reference point are not legitimately different by more than Pi (in which case the algorithm can't be used anyway), no rounding ambiguity should occur.