NR.m: Don't use the sum of the BH masses to normalise anything
In order for NRMMA to provide a consistent framework, it must be clear
what units every quantity is in.  In the past, we have divided some
quantities, such as the initial spin or the initial orbital angular
momentum, by M or M^2, where M (the sum of the initial BH puncture ADM
masses) is the usual unit that we want to use in papers.  However, the
waveform or coordinate data is not treated in this way.  This commit
removes all these normalisations, and makes everything consistent.
All data in NRMMA is now represented in the unit system of the
simulation (which we usually arrange to be the one where M1+M2 = 1, so
this change should not affect any simulations where that is true,
which is almost all).  Any normalisation must now be done external to
NRMMA, for example by using the TotalMass[run] function.