RunFiles.m: Fix problem with RunDirectory variable
Previously we set RunDirectory as a private variable from the setting
of Global`RunDirectory when the package was loaded.  If it had no
value, this was OK because it would remain as a symbol until it was
given a value.  But if it had a value and that value was subsequently
changed, you would never see the new value.