GridFunctions.m: Remove ReadVariables
This function does not make sense in this form at this level of the interface.  We probably want a version of ReadVariables which lists the variables available in the run from all providers.  Each provider would do a filename search for variables.  If multiple variables are stored in a file, it would have to scan each such file.