SimView: Deprecate SimView and add SimulationOverview
SimulationOverview provides a SimulationType option which works similarly to the Method option for NDSolve, namely it can be either a string such as "BBH", or a list of the form {"BBH", opt1 -> val1, …}. This should make it possible to extend SimulationOverview to support many different simulation types in the future. This would probably require more development work on SimulationOverview, but the user facing interface would stay the same.