ReadPsi4: Give an error if the requested radius is not available
Previously the nearest radius was returned.  Now we check that the nearest radius is the same as that asked for within a tolerance of 0.01, which is the precision of the radius from the current output formats.  This logic should really be in the separate output format providers, and we might want to check additional metadata from the HDF5 files to get the "true" double precision radius, rather than relying on the HDF5 dataset name.

Also add a unit test for this problem.