Waveforms.m: Change resampling in new extrapolation code
The previous version was nondeterministic as it used the spacing to determine which DataTable to resample onto.  If the spacings were only different at the level of roundoff, this could lead to unpredictable results.  Instead, we now resample onto the first DataTable, which is usually the lowest extraction radius.  This changes numerical results for users of the new extrapolation interface (only).