DataRegion: Improved handling of attributes.
All attribute names are now strings and the only required ones are "Origin",
"Spacing" and "VariableName". This makes the "Time" attribute optional.

This system is now extensible in the sense that arbitrary attributes can be
given to ToDataTable using the "Attributes" option, which should be a list
of "Name" -> value items.