DataTable: Allow values of Indeterminate and ComplexInfinity
Modify validQ in such a way that it allows Indeterminate and
ComplexInfinity as values for the data in a DataTable, but not for the
coordinates. The implementation is such that we always avoid unpacking
the coordinates array.

This will cause the data part of a DataTable to be unpacked, so
it will be less efficient in terms of storage and performance.