CarpetIOHDF5.m: List maps from all files in a multi-file set
Previously, only maps present in the first file of a file_*.h5 set were being returned.  This should probably be fixed in a more general way, as the same problem could theoretically happen for refinement levels, but with the current implementation in Carpet, this is not possible, as all refinement levels exist on all processes.