Movie.m: Add a framework for working with "frame sets"
A frame set is a set of graphics objects which might be expensive to
compute.  These functions allow you to create a frame set object
(attached to a symbol) and compute frames on demand, which are then
cached.  The entire frameset object can be stored to disk and loaded
back, and the frame set can be Manipulated or rendered as a movie.