CarpetIOHDF5.m: Fix FindGridFunctions when used with h5mma
h5mma's default "Turbo" mode lists only the top-level datasets in an
HDF5 file.  The carpetIOHDF5FileQ function was looking for a dataset
inside a group, and h5mma does not return this.  This caused the
FindGridFunctions test to fail when used with h5mma, though it should
have passed when used without.  The simple fix implemented in this
commit is to search just for the group name, not the dataset, as that is
enough to identify the file as a CarpetIOHDF5 file.