Trackers: Optimise case where two trackers of the same type are used
In this case, it can be faster to read the whole lot together and then
compute the relative vector afterwards.

This commit also generalises the two-tracker versions of the tracker
functions so that the first tracker can have a either a list of indices
or a single index. The list version computes all vectors relative to the
origin specified by the second tracker.